LMS_Project-main/requirements.txt
platformdirs
psycopg2
PyQt5
PyQt5-Qt5
PyQt5-stubs
PyQt5_sip
python-dotenv